+++ src/client/src/components/form/formResultTableDisplayStyles.ts
@@ -1,0 +1,22 @@
+import React from "react";
+
+export const parameterTableHeaderStyles = {
+  "& .MuiFormControl-root": {
+    minWidth: "100%",
+    maxWidth: "100%",
+  },
+  pr: "3px",
+  pl: "3px",
+  maxWidth: "200px",
+  minWidth: "200px",
+};
+
+export const tableCellStyles: React.CSSProperties = {
+  width: "20%",
+  maxWidth: "20%",
+};
+
+export const tableHeaderStyles: React.CSSProperties = {
+  width: "20%",
+  maxWidth: "20%",
+};