+++ src/client/src/pages/detail/form/hydrogeology/hydrotestDisplay.jsx
@@ -4,6 +4,7 @@ import { deleteHydrotest, useDomains } from "../../../../api/fetchApiV2.js";
 import DataDisplayCard from "../../../../components/dataCard/dataDisplayCard.jsx";
 import { FormContainer, FormDisplay, FormValueType } from "../../../../components/form/form";
 import { FormResultTableDisplay } from "../../../../components/form/formResultTableDisplay";
+import { parameterTableHeaderStyles } from "../../../../components/form/formResultTableDisplayStyles";
 import ObservationDisplay from "./observationDisplay.tsx";
 import { getHydrotestParameterUnits } from "./parameterUnits";
 
@@ -41,17 +42,10 @@ const HydrotestDisplay = props => {
                 scope="row"
                 sx={{
                   ...styles,
-                  "& .MuiFormControl-root": {
-                    minWidth: "100%",
-                    maxWidth: "100%",
-                  },
-                  pr: "3px",
-                  pl: "3px",
-                  maxWidth: "200px",
-                  minWidth: "200px",
+                  ...parameterTableHeaderStyles,
                 }}
                 data-cy={`hydrotestResult.${index}.parameter-formDisplay`}>
-                {domains?.data?.find(d => d.id === result.parameterId)?.[i18n.language] || ""}
+                {domains?.data?.find(d => d.id === result.parameterId)?.[i18n.language] ?? ""}
               </TableCell>
               <TableCell sx={styles} data-cy={`hydrotestResult.${index}.value-formDisplay`}>
                 {result?.value && (